requirements.txt
# Required requirements
httpx
motor
pyyaml

# Optional requirements
psutil

# Development requirements
coverage
flake8
flake8-quotes
mongomock_motor
pytest
pytest-asyncio
recommonmark
watchdog
